+++ previous-versions/common/RELEASES.md
@@ -661,3 +661,13 @@ Based on:
 - [typescript v0.1.1] previous-versions/common
 ### Releases
 - [NPM v0.1.1] https://www.npmjs.com/package/@codat/common/v/0.1.1 - previous-versions/common
+
+## 2023-09-13 00:03:45
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.2] previous-versions/common
+### Releases
+- [NPM v0.1.2] https://www.npmjs.com/package/@codat/common/v/0.1.2 - previous-versions/common